mainnet helpers
